+++ media/lua/client/CHC_main.lua
@@ -43,6 +43,7 @@ CHC_main.processOneItem = function(item)
 	local fullType = item:getFullName()
 	local invItem = instanceItem(fullType)
 	local itemDisplayCategory = invItem:getDisplayCategory()
+
 	if not CHC_main.items[fullType] then
 		local toinsert = {
 			item = invItem,
@@ -89,9 +90,10 @@ CHC_main.loadAllItems = function(am)
 	print("Loading items...")
 	for i = 0, amount do
 		local item = allItems:get(i)
-		CHC_main.processOneItem(item)
-
-		nbItems = nbItems + 1
+		if not item:getObsolete() then -- and not item:isHidden() then
+			CHC_main.processOneItem(item)
+			nbItems = nbItems + 1
+		end
 	end
 	showTime(now, "All Items")
 	print(nbItems .. ' items loaded.')