+++ media/lua/client/UI/components/CHC_search_bar.lua
@@ -51,10 +51,10 @@ function CHC_search_bar:create()
     self.searchBar:setText("")
     self.searchBar:setClearButton(true)
     self.searchBarLastText = self.searchBar:getInternalText()
-    -- self.searchBar:setUIName("chat text entry") -- some hardcoded magic
     self.searchBarText = self.searchBarLastText
     self.searchBar.onTextChange = self.onTextChange
     self.searchBar.onOtherKey = CHC_search_bar.onOtherKey
+    -- self.searchBar.onRightMouseDown = self.onRightMouseDown
 
     self:addChild(self.searchBtn)
     self:addChild(self.searchBar)
@@ -79,6 +79,13 @@ function CHC_search_bar:onOtherKey(key)
         self:unfocus()
     end
 end
+
+-- function CHC_search_bar:onRightMouseDown()
+--     local s = self.parent
+--     if s.onRightMouseDownSB then
+--         s.onRightMouseDownSB(s.parent)
+--     end
+-- end
 
 function CHC_search_bar:updateSearchBarLastText()
     local txt = self.searchBar:getInternalText()
@@ -145,6 +152,7 @@ function CHC_search_bar:new(x, y, width, height, searchBarTooltip, onTextChange,
     o.h = height
     o.searchBtnOnClickText = searchBtnOnClickText
     o.onTextChangeSB = onTextChange
+    -- o.onRightMouseDownSB = onRightMouseDown
     o.searchBarTooltip = searchBarTooltip or string.sub(getText("IGUI_CraftUI_Name_Filter"), 1, -2)
     return o
 end